feat: Add excel-to-csv to .gitignore and update build configurations
The .gitignore file has been updated with the addition of 'excel-to-csv'. This ensures that this specific file is ignored by Git, preventing it from being tracked.

In the build configuration, a new property 'dts' has been added under 'rollup' with compiler options set to not emit on error. This allows the TypeScript compiler to continue compiling even if errors are found.

Also, in tsconfig.json, "noEmitOnError" option has been set to false. This change allows TypeScript to compile code even when there are errors present.