chore: Update .gitignore file
Expanded the .gitignore file to include more types of files and directories that should be ignored by Git. This includes logs, diagnostic reports, runtime data, coverage directories, dependency directories, cache directories, build outputs and more. The aim is to prevent unnecessary files from being tracked which are not needed in the repository or for the application to run.